+++ Source/profileGUI.py
@@ -7,6 +7,8 @@ import searchProfile
 import gui
 import filter
 from typing import List
+from typing import Tuple
+from typing import Any
 
 # Class for the buttons to extend
 class SaveSearchBtn(xbmcgui.ControlButton):
@@ -36,6 +38,7 @@ class PlayOneBtn(SaveSearchBtn):
             del poWindow
         else:
             xbmcgui.Dialog().ok("Error", "Your list is empty!")
+        return 1
 
 
 # Class for the Loop Play Button
@@ -51,6 +54,7 @@ class LoopPlayBtn(SaveSearchBtn):
             del lpWindow
         else:
             xbmcgui.Dialog().ok("Error", "Your list is empty!")
+        return 1
 
 # Class for Show List Button
 class ShowListBtn(SaveSearchBtn):
@@ -65,15 +69,18 @@ class ShowListBtn(SaveSearchBtn):
             del slWindow
         else:
             xbmcgui.Dialog().ok("Error", "Your list is empty!")
+        return 1
 
-# Class for Edit Profile Button
-class EditProfileBtn(SaveSearchBtn):
+# Class for Delete Profile Button
+class DeleteProfileBtn(SaveSearchBtn):
     
     def actionTaken(self):
-        print(f'Edit Profile {self.getProfileName()}')
+        searchProfile.removeProfile(self.getProfileName())
+        return 0
 class SavedSearch(xbmcgui.Window):
 
     controlList: List[SaveSearchBtn]
+    controls: List[xbmcgui.Control]
 
     # constructor for window
     def __init__(self):
@@ -86,37 +93,48 @@ class SavedSearch(xbmcgui.Window):
 
         # build inherited object
         super().__init__()
+        self._loadWindow(profiles)
         
+    
+    def _loadWindow(self, profiles: List[Tuple[str, Any]]):
         addon = xbmcaddon.Addon()
         imagesFolder = addon.getAddonInfo('path') + "/images/"
 
         # window specific info
-        controls: List(xbmcgui.Control) = []
         self.controlList = []
+        self.controls = []
+        
         v: int = 200
         for x in profiles:
-            controls.append(xbmcgui.ControlLabel(100, v-75, 400, 50, x[0]))
-            editBtn = EditProfileBtn(200, v, 150, 75, 'Edit Profile', focusTexture=f'{imagesFolder}circle.png', noFocusTexture=f'{imagesFolder}circle.png')
+            self.controls.append(xbmcgui.ControlLabel(100, v-75, 400, 50, x[0]))
+            editBtn = DeleteProfileBtn(200, v, 150, 75, 'Delete Search', focusTexture=f'{imagesFolder}circle.png', noFocusTexture=f'{imagesFolder}circle.png')
             editBtn.profileName = x[0]
-            controls.append(editBtn)
+            self.controls.append(editBtn)
             self.controlList.append(editBtn)
             poBtn = PlayOneBtn(400, v, 150, 75, 'Play One', focusTexture=f'{imagesFolder}circle.png', noFocusTexture=f'{imagesFolder}circle.png')
             poBtn.profileName = x[0]
-            controls.append(poBtn)
+            self.controls.append(poBtn)
             self.controlList.append(poBtn)
             lpBtn = LoopPlayBtn(600, v, 150, 75, 'Loop Play', focusTexture=f'{imagesFolder}circle.png', noFocusTexture=f'{imagesFolder}circle.png')
             lpBtn.profileName = x[0]
-            controls.append(lpBtn)
+            self.controls.append(lpBtn)
             self.controlList.append(lpBtn)
             slBtn = ShowListBtn(800, v, 150, 75, 'Show List', focusTexture=f'{imagesFolder}circle.png', noFocusTexture=f'{imagesFolder}circle.png')
             slBtn.profileName = x[0]
-            controls.append(slBtn)
+            self.controls.append(slBtn)
             self.controlList.append(slBtn)
             v+=200
-        self.addControls(controls)
+        self.addControls(self.controls)
 
     def onControl(self, control: xbmcgui.Control):
         for x in self.controlList:
             if x.getId() == control.getId():
-                x.actionTaken()
-        print(control.getLabel())
+                res = x.actionTaken()
+                if res == 0:
+                    profiles = searchProfile.getAllProfiles()
+                    if (profiles == []):
+                        # show the search options settings window instead
+                        gui.showGui()
+                    else:
+                        self.removeControls(self.controls)
+                        self._loadWindow(profiles)